move simplex.h to be a public header
