+++ src/ccd/support.h
